student-coach/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic-settings==2.3.4
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
APScheduler==3.10.4
bcrypt==3.2.0
